feat: (hopefully) make allocator thread-safe
The location of Heaps need to be stable as Segments contain a pointer to
the owning Heap. This means an Allocator cannot move Heaps once they are
initialised. A std.Segmented list is used as an easy solution.

The atomic operations may be overly strict of even incorrect and will
require careful auditing in the future.