feat!: remove track_allocations and memory_limit options
Removing these options greatly simplifies the code - they can be added
back in the future if desired. As a consequence of removing these
options, libzimalloc has been reworked - it did not actually require
track_allocations to get the size of allocation. This change makes the
redis and lua mimalloc-bench tests succeed.