refactor!: make thread heap list a linked list
In order to permit removing threads from the list while maintaining
stable pointers to heaps a linked list is required. A future
enchancement could replace the use of `std.heap.MemoryPool` with a
custom allocation strategy, since the arena allocator backing a memory
pool does not currently take advantage of the fact that the page
allocator always allocates whole pages.